modify files' and folders' name
