+++ src/main/java/com/nhom10/pbl/controller/AdminController.java
@@ -16,8 +16,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import com.nhom10.pbl.models.Status;
 import com.nhom10.pbl.payload.response.ArticleResponse;
 import com.nhom10.pbl.payload.response.UserResponse;
-import com.nhom10.pbl.payload.resquest.ArticleRequest;
-import com.nhom10.pbl.payload.resquest.UserDTO;
+import com.nhom10.pbl.payload.request.ArticleRequest;
+import com.nhom10.pbl.payload.request.UserDTO;
 import com.nhom10.pbl.security.service.UserService;
 import com.nhom10.pbl.services.ArticleService;
 import com.nhom10.pbl.services.ScheduleServices;
