+++ src/main/java/com/nhom10/pbl/controller/appointment_controller/schedule_controller/scheduleController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.nhom10.pbl.payload.response.ScheduleRespone;
-import com.nhom10.pbl.payload.resquest.ScheduleRequest;
+import com.nhom10.pbl.payload.request.ScheduleRequest;
 import com.nhom10.pbl.services.ScheduleServices;
 
 import jakarta.servlet.http.HttpServletRequest;
