+++ src/main/java/com/nhom10/pbl/payload/request/ArticleRequest.java
@@ -1,4 +1,4 @@
-package com.nhom10.pbl.payload.resquest;
+package com.nhom10.pbl.payload.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
