+++ src/main/java/com/nhom10/pbl/payload/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.nhom10.pbl.payload.resquest;
+package com.nhom10.pbl.payload.request;
 
 import java.sql.Date;
 import lombok.AllArgsConstructor;
