+++ src/main/java/com/nhom10/pbl/security/service/AuthenticateService.java
@@ -15,8 +15,8 @@ import com.nhom10.pbl.models.Role;
 import com.nhom10.pbl.models.UserModel;
 import com.nhom10.pbl.payload.response.AuthenticationResponse;
 import com.nhom10.pbl.payload.response.UserResponse;
-import com.nhom10.pbl.payload.resquest.AuthenticationRequest;
-import com.nhom10.pbl.payload.resquest.RegisterRequest;
+import com.nhom10.pbl.payload.request.AuthenticationRequest;
+import com.nhom10.pbl.payload.request.RegisterRequest;
 import com.nhom10.pbl.repository.RoleRepository;
 import com.nhom10.pbl.repository.UserRepository;
 import com.nhom10.pbl.security.jwt.JWTService;
